+++ packages/labs/analyzer/src/lib/package-analyzer.ts
